CHAT-11 Update user login service and refactor folders.
