+++ src/app/auth/login/login.component.html
@@ -16,7 +16,7 @@
       </div>
     </div>
     <div class="w-full mt-12 px-12 sm:px-44 lg:px-32 xl:px-44">
-      <form [formGroup]="loginForm" (ngSubmit)="onSubmit()" class="flex flex-col text-sm">
+      <form [formGroup]="loginForm" (ngSubmit)="onLogin()" class="flex flex-col text-sm">
         <input type="email" placeholder="Email" formControlName="email"
                [ngClass]="!loginForm.get('email')?.valid && loginForm.get('email')?.touched ? 'border-accent-error' : 'border-accent-gray'"
                class="appearance-none focus:outline-none block border-b bg-transparent w-full p-3 mb-4 text-white placeholder-white"
