--- src/app/auth/models/auth.model.ts
@@ -1,6 +1,0 @@
-export interface AuthModel {
-  email: string;
-  username: string;
-  password: string;
-  confirmPassword: string;
-}
