+++ src/app/models/auth.model.ts
@@ -1,0 +1,11 @@
+export interface UserCreate {
+  email: string;
+  username: string;
+  password: string;
+  confirmPassword: string;
+}
+
+export interface UserLogin {
+  email: string;
+  password: string;
+}
