+++ src/app/validators/auth.validator.ts
@@ -1,6 +1,6 @@
 import {AbstractControl, FormGroup, ValidationErrors, ValidatorFn} from '@angular/forms';
 
-export function passwordValidator(controlPassword: string, controlConfirmPassword: string): ValidatorFn {
+export function authValidator(controlPassword: string, controlConfirmPassword: string): ValidatorFn {
   return (control: AbstractControl): ValidationErrors | null => {
     const formGroup = control as FormGroup;
     const password = formGroup.get(controlPassword)?.value;
