+++ src/environments/environment.dev.ts
@@ -2,6 +2,6 @@ export const environment = {
   production: false,
   hostURL: 'http://localhost',
   port: 4200,
-  apiURL: 'http://localhost:3000/',
+  apiURL: 'http://localhost:3000/api/',
   envName: 'develop'
 };
